refactor(views): make a package out of a single module
